--- frontend/src/views/api-test/components/requestComposition/response.vue
@@ -1,411 +1,0 @@
-<template>
-  <div class="flex h-full min-w-[300px] flex-col">
-    <div
-      class="flex flex-wrap items-center justify-between gap-[8px] border-b border-[var(--color-text-n8)] p-[8px_16px]"
-    >
-      <div class="flex items-center justify-between">
-        <template v-if="props.activeLayout === 'vertical'">
-          <MsButton
-            v-if="props.isExpanded"
-            type="icon"
-            class="!mr-0 !rounded-full bg-[rgb(var(--primary-1))]"
-            size="small"
-            @click="emit('changeExpand', false)"
-          >
-            <icon-down :size="8" />
-          </MsButton>
-          <MsButton
-            v-else
-            type="icon"
-            status="secondary"
-            class="!mr-0 !rounded-full"
-            size="small"
-            @click="emit('changeExpand', true)"
-          >
-            <icon-right :size="8" />
-          </MsButton>
-        </template>
-        <div class="ml-[4px] mr-[24px] font-medium">{{ t('apiTestDebug.responseContent') }}</div>
-        <a-radio-group
-          v-if="!props.hideLayoutSwitch"
-          v-model:model-value="innerLayout"
-          type="button"
-          size="small"
-          @change="(val) => emit('changeLayout', val as Direction)"
-        >
-          <a-radio value="vertical">{{ t('apiTestDebug.vertical') }}</a-radio>
-          <a-radio value="horizontal">{{ t('apiTestDebug.horizontal') }}</a-radio>
-        </a-radio-group>
-      </div>
-      <div
-        v-if="props.response.requestResults[0]?.responseResult?.responseCode"
-        class="flex items-center justify-between gap-[24px]"
-      >
-        <a-popover position="left" content-class="response-popover-content">
-          <div :style="{ color: statusCodeColor }">
-            {{ props.response.requestResults[0].responseResult.responseCode }}
-          </div>
-          <template #content>
-            <div class="flex items-center gap-[8px] text-[14px]">
-              <div class="text-[var(--color-text-4)]">{{ t('apiTestDebug.statusCode') }}</div>
-              <div :style="{ color: statusCodeColor }">
-                {{ props.response.requestResults[0].responseResult.responseCode }}
-              </div>
-            </div>
-          </template>
-        </a-popover>
-        <a-popover position="left" content-class="w-[400px]">
-          <div class="one-line-text text-[rgb(var(--success-7))]">
-            {{ props.response.requestResults[0].responseResult.responseTime }} ms
-          </div>
-          <template #content>
-            <div class="mb-[8px] flex items-center gap-[8px] text-[14px]">
-              <div class="text-[var(--color-text-4)]">{{ t('apiTestDebug.responseTime') }}</div>
-              <div class="text-[rgb(var(--success-7))]">
-                {{ props.response.requestResults[0].responseResult.responseTime }} ms
-              </div>
-            </div>
-            <responseTimeLine :response-timing="timingInfo" />
-          </template>
-        </a-popover>
-        <a-popover position="left" content-class="response-popover-content">
-          <div class="one-line-text text-[rgb(var(--success-7))]">
-            {{ props.response.requestResults[0].responseResult.responseSize }} bytes
-          </div>
-          <template #content>
-            <div class="flex items-center gap-[8px] text-[14px]">
-              <div class="text-[var(--color-text-4)]">{{ t('apiTestDebug.responseSize') }}</div>
-              <div class="one-line-text text-[rgb(var(--success-7))]">
-                {{ props.response.requestResults[0].responseResult.responseSize }} bytes
-              </div>
-            </div>
-          </template>
-        </a-popover>
-        <!-- <a-popover position="left" content-class="response-popover-content">
-          <div class="text-[var(--color-text-1)]">{{ props.response.env }}</div>
-          <template #content>
-            <div class="flex items-center gap-[8px] text-[14px]">
-              <div class="text-[var(--color-text-4)]">{{ t('apiTestDebug.runningEnv') }}</div>
-              <div class="text-[var(--color-text-1)]">{{ props.response.env }}</div>
-            </div>
-          </template>
-        </a-popover>
-        <a-popover position="left" content-class="response-popover-content">
-          <div class="text-[var(--color-text-1)]">{{ props.response.resource }}</div>
-          <template #content>
-            <div class="flex items-center gap-[8px] text-[14px]">
-              <div class="text-[var(--color-text-4)]">{{ t('apiTestDebug.resourcePool') }}</div>
-              <div class="text-[var(--color-text-1)]">{{ props.response.resource }}</div>
-            </div>
-          </template>
-        </a-popover> -->
-      </div>
-    </div>
-    <a-spin :loading="props.loading" class="h-[calc(100%-42px)] w-full px-[18px] pb-[18px]">
-      <a-tabs v-model:active-key="activeTab" class="no-content border-b border-[var(--color-text-n8)]">
-        <a-tab-pane v-for="item of responseTabList" :key="item.value" :title="item.label" />
-      </a-tabs>
-      <div class="response-container">
-        <MsCodeEditor
-          v-if="activeTab === ResponseComposition.BODY"
-          ref="responseEditorRef"
-          :model-value="props.response.requestResults[0]?.responseResult?.body"
-          :language="responseLanguage"
-          theme="vs"
-          height="100%"
-          :languages="[LanguageEnum.JSON, LanguageEnum.HTML, LanguageEnum.XML, LanguageEnum.PLAINTEXT]"
-          :show-full-screen="false"
-          :show-theme-change="false"
-          show-language-change
-          show-charset-change
-          read-only
-        >
-          <template #rightTitle>
-            <a-button type="outline" class="arco-btn-outline--secondary p-[0_8px]" size="mini" @click="copyScript">
-              <template #icon>
-                <MsIcon type="icon-icon_copy_outlined" class="text-var(--color-text-4)" size="12" />
-              </template>
-            </a-button>
-          </template>
-        </MsCodeEditor>
-        <MsCodeEditor
-          v-else-if="activeTab === ResponseComposition.CONSOLE"
-          :model-value="response.console.trim()"
-          :language="LanguageEnum.PLAINTEXT"
-          theme="MS-text"
-          height="100%"
-          :show-full-screen="false"
-          :show-theme-change="false"
-          :show-language-change="false"
-          :show-charset-change="false"
-          read-only
-        >
-        </MsCodeEditor>
-        <div
-          v-else-if="
-            activeTab === ResponseComposition.HEADER ||
-            activeTab === ResponseComposition.REAL_REQUEST ||
-            activeTab === ResponseComposition.EXTRACT
-          "
-          class="h-full rounded-[var(--border-radius-small)] bg-[var(--color-text-n9)] p-[12px]"
-        >
-          <pre class="response-header-pre">{{ getResponsePreContent(activeTab) }}</pre>
-        </div>
-        <MsBaseTable v-else-if="activeTab === 'ASSERTION'" v-bind="propsRes" v-on="propsEvent">
-          <template #status="{ record }">
-            <MsTag :type="record.status === 1 ? 'success' : 'danger'" theme="light">
-              {{ record.status === 1 ? t('common.success') : t('common.fail') }}
-            </MsTag>
-          </template>
-        </MsBaseTable>
-      </div>
-    </a-spin>
-  </div>
-</template>
-
-<script setup lang="ts">
-  import { useClipboard, useVModel } from '@vueuse/core';
-  import { Message } from '@arco-design/web-vue';
-
-  import MsButton from '@/components/pure/ms-button/index.vue';
-  import MsCodeEditor from '@/components/pure/ms-code-editor/index.vue';
-  import { LanguageEnum } from '@/components/pure/ms-code-editor/types';
-  import MsIcon from '@/components/pure/ms-icon-font/index.vue';
-  import type { Direction } from '@/components/pure/ms-split-box/index.vue';
-  import MsBaseTable from '@/components/pure/ms-table/base-table.vue';
-  import { MsTableColumn } from '@/components/pure/ms-table/type';
-  import useTable from '@/components/pure/ms-table/useTable';
-  import MsTag from '@/components/pure/ms-tag/ms-tag.vue';
-  import responseTimeLine from '@/views/api-test/components/responseTimeLine.vue';
-
-  import { useI18n } from '@/hooks/useI18n';
-
-  import { ResponseComposition } from '@/enums/apiEnum';
-
-  import type { RequestParam } from './index.vue';
-
-  export interface Response {
-    requestResults: {
-      body: string;
-      headers: string;
-      responseResult: {
-        body: string;
-        contentType: string;
-        headers: string;
-        dnsLookupTime: number;
-        downloadTime: number;
-        latency: number;
-        responseCode: number;
-        responseTime: number;
-        responseSize: number;
-        socketInitTime: number;
-        sslHandshakeTime: number;
-        tcpHandshakeTime: number;
-        transferStartTime: number;
-      };
-    }[]; // 请求结果
-    console: string;
-  }
-
-  const props = withDefaults(
-    defineProps<{
-      activeTab: keyof typeof ResponseComposition;
-      activeLayout?: Direction;
-      isExpanded: boolean;
-      response: Response;
-      request?: RequestParam;
-      hideLayoutSwitch?: boolean; // 隐藏布局切换
-      loading?: boolean;
-    }>(),
-    {
-      activeLayout: 'vertical',
-      hideLayoutSwitch: false,
-    }
-  );
-  const emit = defineEmits<{
-    (e: 'update:activeLayout', value: Direction): void;
-    (e: 'update:activeTab', value: keyof typeof ResponseComposition): void;
-    (e: 'changeExpand', value: boolean): void;
-    (e: 'changeLayout', value: Direction): void;
-  }>();
-
-  const { t } = useI18n();
-
-  const innerLayout = useVModel(props, 'activeLayout', emit);
-  const activeTab = useVModel(props, 'activeTab', emit);
-  // 响应时间信息
-  const timingInfo = computed(() => {
-    const {
-      dnsLookupTime,
-      downloadTime,
-      latency,
-      responseTime,
-      socketInitTime,
-      sslHandshakeTime,
-      tcpHandshakeTime,
-      transferStartTime,
-    } = props.response.requestResults[0].responseResult;
-    return {
-      dnsLookupTime,
-      tcpHandshakeTime,
-      sslHandshakeTime,
-      socketInitTime,
-      latency,
-      downloadTime,
-      transferStartTime,
-      responseTime,
-    };
-  });
-  // 响应状态码对应颜色
-  const statusCodeColor = computed(() => {
-    const code = props.response.requestResults[0].responseResult.responseCode;
-    if (code >= 200 && code < 300) {
-      return 'rgb(var(--success-7)';
-    }
-    if (code >= 300 && code < 400) {
-      return 'rgb(var(--warning-7)';
-    }
-    return 'rgb(var(--danger-7)';
-  });
-  // 响应体语言类型
-  const responseLanguage = computed(() => {
-    const { contentType } = props.response.requestResults[0].responseResult;
-    if (contentType.includes('json')) {
-      return LanguageEnum.JSON;
-    }
-    if (contentType.includes('html')) {
-      return LanguageEnum.HTML;
-    }
-    if (contentType.includes('xml')) {
-      return LanguageEnum.XML;
-    }
-    return LanguageEnum.PLAINTEXT;
-  });
-  const responseEditorRef = ref<InstanceType<typeof MsCodeEditor>>();
-
-  const responseTabList = [
-    {
-      label: t('apiTestDebug.responseBody'),
-      value: ResponseComposition.BODY,
-    },
-    {
-      label: t('apiTestDebug.responseHeader'),
-      value: ResponseComposition.HEADER,
-    },
-    {
-      label: t('apiTestDebug.realRequest'),
-      value: ResponseComposition.REAL_REQUEST,
-    },
-    {
-      label: t('apiTestDebug.console'),
-      value: ResponseComposition.CONSOLE,
-    },
-    // {
-    //   label: t('apiTestDebug.extract'),
-    //   value: ResponseComposition.EXTRACT,
-    // },
-    // {
-    //   label: t('apiTestDebug.assertion'),
-    //   value: ResponseComposition.ASSERTION,
-    // }, // TODO:断言暂时没加
-  ];
-
-  const { copy, isSupported } = useClipboard();
-
-  function copyScript() {
-    if (isSupported) {
-      copy(props.response.requestResults[0].responseResult.body);
-      Message.success(t('common.copySuccess'));
-    } else {
-      Message.warning(t('apiTestDebug.copyNotSupport'));
-    }
-  }
-
-  function getResponsePreContent(type: keyof typeof ResponseComposition) {
-    switch (type) {
-      case ResponseComposition.HEADER:
-        return props.response.requestResults[0].responseResult.headers.trim();
-      case ResponseComposition.REAL_REQUEST:
-        return props.response.requestResults[0].body
-          ? `${t('apiTestDebug.requestUrl')}:\n${props.request?.url}\n${t('apiTestDebug.header')}:\n${
-              props.response.requestResults[0].headers
-            }\nBody:\n${props.response.requestResults[0].body.trim()}`
-          : '';
-      // case ResponseComposition.EXTRACT:
-      //   return Object.keys(props.response.extract)
-      //     .map((e) => `${e}: ${props.response.extract[e]}`)
-      //     .join('\n'); // TODO:断言暂时没加
-      default:
-        return '';
-    }
-  }
-
-  const columns: MsTableColumn = [
-    {
-      title: 'apiTestDebug.content',
-      dataIndex: 'content',
-      showTooltip: true,
-    },
-    {
-      title: 'apiTestDebug.status',
-      dataIndex: 'status',
-      slotName: 'status',
-      width: 80,
-    },
-    {
-      title: '',
-      dataIndex: 'desc',
-      showTooltip: true,
-    },
-  ];
-  const { propsRes, propsEvent } = useTable(() => Promise.resolve([]), {
-    scroll: { x: '100%' },
-    columns,
-  });
-  propsRes.value.data = [
-    {
-      id: new Date().getTime(),
-      content: 'Response Code equals: 200',
-      status: 1,
-      desc: '',
-    },
-    {
-      id: new Date().getTime(),
-      content: '$.users[1].age REGEX: 31',
-      status: 0,
-      desc: `Value expected to match regexp '31', but it did not match: '30' match: '30'`,
-    },
-  ] as any;
-</script>
-
-<style lang="less">
-  .response-popover-content {
-    padding: 4px 8px;
-    .arco-popover-content {
-      @apply mt-0;
-
-      font-size: 14px;
-      line-height: 22px;
-    }
-  }
-</style>
-
-<style lang="less" scoped>
-  .response-container {
-    margin-top: 8px;
-    height: calc(100% - 48px);
-    .response-header-pre {
-      @apply h-full overflow-auto  bg-white;
-      .ms-scroll-bar();
-
-      padding: 8px 12px;
-      border-radius: var(--border-radius-small);
-    }
-  }
-  :deep(.arco-table-th) {
-    background-color: var(--color-text-n9);
-  }
-  :deep(.arco-tabs-tab) {
-    @apply leading-none;
-  }
-</style>
