+++ frontend/src/views/api-test/components/requestComposition/response/edit.vue
@@ -1,0 +1,7 @@
+<template>
+  <div> </div>
+</template>
+
+<script setup lang="ts"></script>
+
+<style lang="less" scoped></style>
